# lib/api-zod/src/generated/types/registerInput.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Api
 * AetherPulse API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterInput {
  /**
   * @minLength 3
   * @maxLength 20
   */
  username: string
  /** @minLength 6 */
  password: string
  /**
   * @minLength 1
   * @maxLength 32
   */
  displayName: string
}