# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Api
 * AetherPulse API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserStatus } from './userStatus'

export interface User {
  id: number
  username: string
  displayName: string
  /** @nullable */
  avatarUrl?: string | null
  /** @nullable */
  bannerUrl?: string | null
  /** @nullable */
  bio?: string | null
  /** @nullable */
  pronouns?: string | null
  /** @nullable */
  website?: string | null
  /** @nullable */
  location?: string | null
  status: UserStatus
  /** @nullable */
  customStatus?: string | null
  /** @nullable */
  accentColor?: string | null
  /** @nullable */
  primaryColor?: string | null
  /** @nullable */
  displayNameStyle?: string | null
  /** @nullable */
  profileGradient?: string | null
  badges?: string[]
  createdAt: string
}