# shared/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * Api
 * AetherPulse API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string
}

export interface RegisterInput {
  /**
   * @minLength 3
   * @maxLength 20
   */
  username: string
  /** @minLength 6 */
  password: string
  /**
   * @minLength 1
   * @maxLength 32
   */
  displayName: string
}

export interface LoginInput {
  username: string
  password: string
}

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]

export const UserStatus = {
  online: 'online',
  away: 'away',
  busy: 'busy',
  offline: 'offline',
} as const

export interface User {
  id: number
  username: string
  displayName: string
  /** @nullable */
  avatarUrl?: string | null
  /** @nullable */
  bannerUrl?: string | null
  /** @nullable */
  bio?: string | null
  /** @nullable */
  pronouns?: string | null
  /** @nullable */
  website?: string | null
  /** @nullable */
  location?: string | null
  status: UserStatus
  /** @nullable */
  customStatus?: string | null
  /** @nullable */
  accentColor?: string | null
  /** @nullable */
  primaryColor?: string | null
  /** @nullable */
  displayNameStyle?: string | null
  /** @nullable */
  profileGradient?: string | null
  /** @nullable */
  avatarFrame?: string | null
  /** @nullable */
  profileTheme?: string | null
  /** @nullable */
  customTheme?: string | null
  /** @nullable */
  preferredTheme?: string | null
  badges?: string[]
  createdAt: string
}

export interface AuthResponse {
  user: User
}

export type UserUpdateStatus = (typeof UserUpdateStatus)[keyof typeof UserUpdateStatus]

export const UserUpdateStatus = {
  online: 'online',
  away: 'away',
  busy: 'busy',
  offline: 'offline',
} as const

export interface UserUpdate {
  displayName?: string
  /**
   * @maxLength 190
   * @nullable
   */
  bio?: string | null
  /**
   * @maxLength 40
   * @nullable
   */
  pronouns?: string | null
  /**
   * @maxLength 120
   * @nullable
   */
  website?: string | null
  /**
   * @maxLength 40
   * @nullable
   */
  location?: string | null
  /**
   * @maxLength 128
   * @nullable
   */
  customStatus?: string | null
  /** @pattern ^#[0-9A-Fa-f]{6}$ */
  accentColor?: string
  /** @nullable */
  primaryColor?: string | null
  /** @nullable */
  displayNameStyle?: string | null
  /** @nullable */
  profileGradient?: string | null
  /** @nullable */
  avatarFrame?: string | null
  /** @nullable */
  profileTheme?: string | null
  /** @nullable */
  customTheme?: string | null
  /** @nullable */
  preferredTheme?: string | null
  /** @nullable */
  avatarUrl?: string | null
  /** @nullable */
  bannerUrl?: string | null
  status?: UserUpdateStatus
  badges?: string[]
}

export type RoomQuality = (typeof RoomQuality)[keyof typeof RoomQuality]

export const RoomQuality = {
  '360p': '360p',
  '480p': '480p',
  '720p': '720p',
  '1080p': '1080p',
  '1440p': '1440p',
} as const

export interface Room {
  id: string
  name: string
  inviteCode: string
  ownerId: number
  memberCount: number
  members?: User[]
  quality?: RoomQuality
  isActive?: boolean
  createdAt: string
}

export type RoomInputQuality = (typeof RoomInputQuality)[keyof typeof RoomInputQuality]

export const RoomInputQuality = {
  '360p': '360p',
  '480p': '480p',
  '720p': '720p',
  '1080p': '1080p',
  '1440p': '1440p',
} as const

export interface RoomInput {
  /**
   * @minLength 1
   * @maxLength 64
   */
  name: string
  quality?: RoomInputQuality
}

export interface JoinRoomInput {
  inviteCode: string
}

export type MessageType = (typeof MessageType)[keyof typeof MessageType]

export const MessageType = {
  text: 'text',
  system: 'system',
  file: 'file',
} as const

export interface Message {
  id: number
  roomId: string
  userId: number
  content: string
  type: MessageType
  user?: User
  createdAt: string
}

export type MessageInputType = (typeof MessageInputType)[keyof typeof MessageInputType]

export const MessageInputType = {
  text: 'text',
  file: 'file',
} as const

export interface MessageInput {
  /**
   * @minLength 1
   * @maxLength 4000
   */
  content: string
  type?: MessageInputType
}

export interface ActivityEvent {
  type: string
  description: string
  /** @nullable */
  userId?: number | null
  createdAt: string
}

export interface RoomActivity {
  roomId: string
  memberCount: number
  messageCount: number
  events: ActivityEvent[]
}

export interface RefreshTokenBody {
  refreshToken: string
}