# shared/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AetherPulse API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterInput'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
        '409':
          description: Username already taken
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Not authenticated
  /auth/refresh:
    post:
      operationId: refresh
      tags: [auth]
      summary: Refresh JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenBody'
      responses:
        '200':
          description: Token refresh successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid token
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [username, password, displayName]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 20
        password:
          type: string
          minLength: 6
        displayName:
          type: string
          minLength: 1
          maxLength: 32
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      required: [id, username, displayName, status, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
        avatarUrl:
          type: ['string', 'null']
        bannerUrl:
          type: ['string', 'null']
        bio:
          type: ['string', 'null']
        pronouns:
          type: ['string', 'null']
        website:
          type: ['string', 'null']
        location:
          type: ['string', 'null']
        status:
          type: string
          enum: [online, away, busy, offline]
        customStatus:
          type: ['string', 'null']
        accentColor:
          type: ['string', 'null']
        primaryColor:
          type: ['string', 'null']
        displayNameStyle:
          type: ['string', 'null']
        profileGradient:
          type: ['string', 'null']
        avatarFrame:
          type: ['string', 'null']
        profileTheme:
          type: ['string', 'null']
        customTheme:
          type: ['string', 'null']
        preferredTheme:
          type: ['string', 'null']
        badges:
          type: array
          items:
            type: string
        createdAt:
          type: string
...
    RefreshTokenBody:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string